pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ap-invoice-env"
version = "1.0.0"
description = "Real-world Accounts Payable Invoice Processor - OpenEnv environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "openenv-core>=0.2.1",
]

[project.optional-dependencies]
dev = ["openai>=1.35.0"]

[tool.uv]
dev-dependencies = ["openai>=1.35.0"]

[tool.setuptools]
py-modules = ["client", "inference", "models"]

[tool.setuptools.packages.find]
include = ["server*"]
